Removed Twig

+++ src/Service/Type.php
@@ -7,7 +7,7 @@ final class Type
     public const TYPE_ARRAY = "array";
     public const TYPE_BOOLEAN = "boolean";
     public const TYPE_DOUBLE = "double";
-    public const TYPE_FLOAT = "double";
+    public const TYPE_FLOAT = "float";
     public const TYPE_INTEGER = "integer";
     public const TYPE_OBJECT = "object";
     public const TYPE_RESOURCE = "resource";
